sysadmin/status.io/requirements.txt
statusio-python
requests
